+++ app/controllers/application_controller.rb
@@ -1,12 +1,6 @@
 # frozen_string_literal: true
 
 class ApplicationController < ActionController::Base
-  rescue_from Notifications::Client::NotFoundError, with: :validation_notice_request_error
-  rescue_from Notifications::Client::ServerError, with: :validation_notice_request_error
-  rescue_from Notifications::Client::RequestError, with: :validation_notice_request_error
-  rescue_from Notifications::Client::ClientError, with: :validation_notice_request_error
-  rescue_from Notifications::Client::BadRequestError, with: :validation_notice_request_error
-
   before_action :find_current_local_authority_from_subdomain
   before_action :prevent_caching
   before_action :set_current_user
@@ -41,13 +35,6 @@ class ApplicationController < ActionController::Base
 
   def disable_flash_header
     @disable_flash_header = true
-  end
-
-  def validation_notice_request_error(exception)
-    flash[:error] = "Notify was unable to send applicant email. Please contact the applicant directly."
-    flash[:notice] = "Document validation successful. Application is ready for assessment."
-    Appsignal.send_error(exception)
-    render "planning_applications/show"
   end
 
   def set_current_user
